+++ database/migrations/2025_04_07_162213_create_permission_role_table.php
@@ -11,10 +11,17 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('permission_role', function (Blueprint $table) {
-            $table->foreignId('permission_id')->constrained()->onDelete('cascade');
-            $table->foreignId('role_id')->constrained()->onDelete('cascade');
-        });
+        if (!Schema::hasTable('permission_role')) {
+            Schema::create('permission_role', function (Blueprint $table) {
+                $table->foreignId('permission_id')
+                      ->constrained()
+                      ->onDelete('cascade');
+
+                $table->foreignId('role_id')
+                      ->constrained()
+                      ->onDelete('cascade');
+            });
+        }
     }
 
     /**
@@ -22,6 +29,8 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::dropIfExists('permission_role');
+        if (Schema::hasTable('permission_role')) {
+            Schema::dropIfExists('permission_role');
+        }
     }
 };
